# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Event Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/events/create": {
            "post": {
                "description": "Accepts EventCreateInputDTO. Same project_id + message_id on retry returns the existing event (idempotent). ip and user_agent are optional fields on the JSON body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Ingest analytics event",
                "parameters": [
                    {
                        "description": "Event payload (see model for per-field descriptions)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateOutputDTO"
                        }
                    }
                }
            }
        },
        "/v1/events/create-batch": {
            "post": {
                "description": "Accepts EventCreateBatchInputDTO. All rows are inserted in one transaction. Same rules as single create; duplicate project_id+message_id within the request body is rejected. Each item may include ip and user_agent on the JSON body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Ingest multiple analytics events (atomic batch)",
                "parameters": [
                    {
                        "description": "Batch payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateBatchInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateBatchOutputDTO"
                        }
                    }
                }
            }
        },
        "/v1/events/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Soft delete event (sets deleted_at)",
                "parameters": [
                    {
                        "description": "Delete Event Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDeleteInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDeleteOutputDTO"
                        }
                    }
                }
            }
        },
        "/v1/events/detail": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get event detail",
                "parameters": [
                    {
                        "description": "Get Event Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDetailInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDetailOutputDTO"
                        }
                    }
                }
            }
        },
        "/v1/events/query": {
            "post": {
                "description": "Filters: all list fields use OR within the field (SQL IN, or ILIKE OR for names). names are substring terms; ips and user_agents are exact. Combined with AND across different fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Returns list of events by query",
                "parameters": [
                    {
                        "description": "Query parameters",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventQueryInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventQueryOutputDTO"
                        }
                    }
                }
            }
        },
        "/v1/events/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Update event",
                "parameters": [
                    {
                        "description": "Update Event Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventUpdateInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventUpdateOutputDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateBatchInputDTO": {
            "type": "object",
            "required": [
                "events"
            ],
            "properties": {
                "events": {
                    "description": "Events is the list of events to ingest in one atomic transaction (max 100).",
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateInputDTO"
                    }
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateBatchOutputDTO": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO"
                    }
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateInputDTO": {
            "type": "object",
            "required": [
                "distinct_id",
                "message_id",
                "name",
                "occurred_at",
                "payload",
                "project_id"
            ],
            "properties": {
                "distinct_id": {
                    "description": "DistinctID is the stable person/device id for analytics (e.g. UUID in localStorage until login).",
                    "type": "string",
                    "example": "anon_6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                },
                "ip": {
                    "description": "IP is the client IP for this event (e.g. from your API gateway or client); omit if unknown.",
                    "type": "string",
                    "example": "203.0.113.10"
                },
                "message_id": {
                    "description": "MessageID is a new UUID per request so retries do not duplicate rows (idempotency key).",
                    "type": "string",
                    "example": "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                },
                "name": {
                    "description": "Name is the event key (snake_case recommended), e.g. signup_started, purchase_completed.",
                    "type": "string",
                    "maxLength": 255,
                    "example": "page_viewed"
                },
                "occurred_at": {
                    "description": "OccurredAt is client-side event time in RFC3339 (used with server received_at for clock skew).",
                    "type": "string",
                    "example": "2026-04-12T10:30:00.123Z"
                },
                "payload": {
                    "description": "Payload is a JSON object of custom properties (strings, numbers, booleans) for this event.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "project_id": {
                    "description": "ProjectID identifies which product or tenant this event belongs to (from API key, config, or backend caller).",
                    "type": "string",
                    "example": "proj_live_01"
                },
                "session_id": {
                    "description": "SessionID ties events to one browser session or visit; omit if not tracking sessions.",
                    "type": "string",
                    "example": "sess_01j8xyz"
                },
                "user_agent": {
                    "description": "UserAgent is the raw user-agent for this event; omit if unknown.",
                    "type": "string",
                    "example": "Mozilla/5.0 (compatible; ...)"
                },
                "user_id": {
                    "description": "UserID is your application’s user id when the visitor is authenticated; omit if anonymous.",
                    "type": "string",
                    "example": "usr_42"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventCreateOutputDTO": {
            "type": "object",
            "properties": {
                "event": {
                    "description": "Event is the stored row including server id, received_at, and request-provided ip / user_agent when present.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO"
                        }
                    ]
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "description": "DeletedAt is set when the row is soft-deleted.",
                    "type": "string"
                },
                "distinct_id": {
                    "description": "DistinctID is the stable analytics identity (e.g. anonymous id in localStorage) for funnels and retention.",
                    "type": "string",
                    "example": "anon_6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                },
                "id": {
                    "description": "ID is the server-generated UUID for this stored row.",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ip": {
                    "description": "IP is the client address stored with the event (caller-supplied, e.g. from edge or client).",
                    "type": "string",
                    "example": "203.0.113.10"
                },
                "message_id": {
                    "description": "MessageID uniquely identifies this delivery; use a new UUID per HTTP request for idempotent retries.",
                    "type": "string",
                    "example": "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                },
                "name": {
                    "description": "Name is the event name / type key (e.g. page_viewed, button_clicked).",
                    "type": "string",
                    "example": "page_viewed"
                },
                "occurred_at": {
                    "description": "OccurredAt is when the action happened on the client (RFC3339); may differ from server time.",
                    "type": "string",
                    "example": "2026-04-12T10:30:00Z"
                },
                "payload": {
                    "description": "Payload holds arbitrary JSON properties for segmentation and reporting.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "project_id": {
                    "description": "ProjectID scopes the event to one product or tenant (from your API key, env, or calling service).",
                    "type": "string",
                    "example": "proj_live_01"
                },
                "received_at": {
                    "description": "ReceivedAt is when this service persisted the event (set by the server).",
                    "type": "string",
                    "example": "2026-04-12T10:30:00.456Z"
                },
                "session_id": {
                    "description": "SessionID groups events within one visit or tab session; optional for server-only events.",
                    "type": "string",
                    "example": "sess_01j8xyz"
                },
                "user_agent": {
                    "description": "UserAgent is the raw user-agent string stored with the event (caller-supplied).",
                    "type": "string",
                    "example": "Mozilla/5.0 ..."
                },
                "user_id": {
                    "description": "UserID is your logged-in account id when available; omitted before authentication.",
                    "type": "string",
                    "example": "usr_42"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventDeleteInputDTO": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventDeleteOutputDTO": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventDetailInputDTO": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventDetailOutputDTO": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventListMetaDTO": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.PaginationOutputParams"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventQueryInputDTO": {
            "type": "object",
            "properties": {
                "distinct_ids": {
                    "description": "DistinctIDs matches any listed distinct_id.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ids": {
                    "description": "IDs filters by event row UUIDs.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ips": {
                    "description": "IPs matches any listed ip (exact).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "message_ids": {
                    "description": "MessageIDs matches any listed message_id (UUID each).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "names": {
                    "description": "Names matches if event name matches any term (substring ILIKE per term).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "project_ids": {
                    "description": "ProjectIDs matches any listed project_id.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_ids": {
                    "description": "SessionIDs matches any listed session_id.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "user_agents": {
                    "description": "UserAgents matches any listed user_agent (exact).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_ids": {
                    "description": "UserIDs matches any listed user_id.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventQueryOutputDTO": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventListMetaDTO"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventUpdateInputDTO": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "session_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.EventUpdateOutputDTO": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/github_com_go-web-services_go-service-event_pkg_client_dto.EventDTO"
                }
            }
        },
        "github_com_go-web-services_go-service-event_pkg_client_dto.PaginationOutputParams": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    }
}